teams_2/requirements.txt
botbuilder-core>=4.15.0
botbuilder-schema>=4.15.0
aiohttp>=3.8.0
langgraph-sdk>=0.1.29
python-dotenv>=0.19.0

